fix ossimNitfTileSource_12 code
